requirements.txt
asgiref==3.8.1
Django==4.2.19
djangorestframework==3.15.2
pytz==2025.1
sqlparse==0.5.3
django-cors-headers==4.6.0
gunicorn==23.0.0
whitenoise==6.8.2

